refactor: split asset utilities into modules
