--- src/main/assets.ts
@@ -1,318 +1,0 @@
-import fs from 'fs';
-import path from 'path';
-import { app as electronApp } from 'electron';
-import type { Protocol, IpcMain } from 'electron';
-import os from 'os';
-import unzipper from 'unzipper';
-import sharp from 'sharp';
-import { readProjectMeta, writeProjectMeta } from './projectMeta';
-import { generatePackIcon } from './icon';
-import {
-  PACK_FORMATS,
-  displayForFormat,
-  PackFormatInfo,
-} from '../shared/packFormat';
-import { toPosixPath } from '../shared/toPosixPath';
-
-/** URL pointing to Mojang's version manifest which lists all official releases. */
-const VERSION_MANIFEST =
-  'https://launchermeta.mojang.com/mc/game/version_manifest.json';
-
-/** Base directory used to store cached client JARs and extracted assets. */
-const basePath = electronApp?.getPath
-  ? electronApp.getPath('userData')
-  : os.tmpdir();
-/** Directory that contains one sub folder per Minecraft version. */
-const cacheDir = path.join(basePath, 'assets-cache');
-
-// Paths used by the custom asset and vanilla protocols. These are updated when
-// a project is opened so the protocols can resolve relative texture URLs.
-let cacheTexturesDir = '';
-// Base path for the currently active project used by the asset protocol.
-let activeProjectDir = '';
-
-/**
- * Fetch a JSON document from the given URL.
- *
- * The request is performed using `fetch` and the body parsed as JSON. When the
- * manifest URL is requested and the network is unavailable, a bundled copy is
- * used as a fallback.
- *
- * @param url URL to request.
- * @returns Parsed JSON typed as `T`.
- * @throws If the request fails.
- */
-async function fetchJson<T>(url: string): Promise<T> {
-  try {
-    const res = await fetch(url);
-    if (!res.ok) throw new Error(`Failed to fetch ${url}`);
-    return res.json() as Promise<T>;
-  } catch (err) {
-    // Fall back to a bundled manifest when offline
-    if (url === VERSION_MANIFEST) {
-      const local = path.join(__dirname, '../shared/version_manifest.json');
-      const data = await fs.promises.readFile(local, 'utf-8');
-      return JSON.parse(data) as T;
-    }
-    throw err;
-  }
-}
-
-/**
- * Download a file from the given URL and write it to disk.
- *
- * The target directory is created if it does not exist. The body is saved as a
- * binary buffer so any file type can be retrieved.
- *
- * @param url URL to download.
- * @param dest Absolute output path on disk.
- * @returns Promise that resolves when the file is written.
- */
-async function downloadFile(url: string, dest: string): Promise<void> {
-  const res = await fetch(url);
-  if (!res.ok) throw new Error(`Failed to download ${url}`);
-  await fs.promises.mkdir(path.dirname(dest), { recursive: true });
-  const array = new Uint8Array(await res.arrayBuffer());
-  await fs.promises.writeFile(dest, array);
-}
-
-/**
- * Extract all files from a zip archive to the given directory.
- *
- * The operation streams the archive to disk to avoid buffering large files in
- * memory.
- *
- * @param src Path to the source zip archive.
- * @param dest Directory that receives the extracted files.
- * @returns Promise that resolves once extraction completes.
- */
-async function extractZip(src: string, dest: string): Promise<void> {
-  await new Promise<void>((resolve, reject) => {
-    fs.createReadStream(src)
-      .pipe(unzipper.Extract({ path: dest }))
-      .on('close', resolve)
-      .on('error', reject);
-  });
-}
-
-/**
- * Ensure that the assets for the given Minecraft version are available locally.
- *
- * Algorithm:
- * 1. Build a version specific cache directory and check if textures already
- *    exist.
- * 2. If present return the path immediately.
- * 3. Otherwise download the corresponding client JAR from Mojang using the
- *    version manifest.
- * 4. Extract the JAR into the cache so individual textures can be accessed.
- *
- * @param version Minecraft version string, e.g. `1.21.1`.
- * @returns Absolute path to the extracted client folder containing `assets/`.
- * @throws If the version cannot be resolved or downloads fail.
- */
-async function ensureAssets(version: string): Promise<string> {
-  const base = path.join(cacheDir, version);
-  const assetsPath = path.join(base, 'client');
-  const texDir = path.join(assetsPath, 'assets', 'minecraft', 'textures');
-  if (fs.existsSync(texDir)) return assetsPath;
-
-  await fs.promises.mkdir(base, { recursive: true });
-  const manifest = await fetchJson<{
-    versions: Array<{ id: string; url: string }>;
-  }>(VERSION_MANIFEST);
-  const entry = manifest.versions.find((v) => v.id === version);
-  if (!entry) throw new Error(`Version ${version} not found`);
-  const ver = await fetchJson<{ downloads: { client: { url: string } } }>(
-    entry.url
-  );
-  const jarUrl = ver.downloads.client.url;
-  const jarPath = path.join(base, 'client.jar');
-  if (!fs.existsSync(jarPath)) await downloadFile(jarUrl, jarPath);
-  await extractZip(jarPath, assetsPath);
-  return assetsPath;
-}
-
-export { fetchJson, downloadFile, ensureAssets };
-
-/**
- * Fetch the official Minecraft version manifest and return the list of
- * available version IDs.
- */
-export async function listPackFormats(): Promise<PackFormatInfo[]> {
-  return PACK_FORMATS.map((r) => ({
-    format: r.format,
-    label: displayForFormat(r.format) as string,
-  }));
-}
-
-/**
- * Recursively list all texture paths available for the given project version.
- */
-export async function listTextures(projectPath: string): Promise<string[]> {
-  const meta = await readProjectMeta(projectPath);
-  const root = await ensureAssets(meta.minecraft_version);
-  const texRoot = path.join(root, 'assets', 'minecraft', 'textures');
-  const out: string[] = [];
-  const walk = async (dir: string) => {
-    const entries = await fs.promises.readdir(dir, { withFileTypes: true });
-    for (const entry of entries) {
-      const p = path.join(dir, entry.name);
-      if (entry.isDirectory()) {
-        await walk(p);
-      } else if (entry.name.endsWith('.png')) {
-        const rel = path.relative(texRoot, p);
-        out.push(toPosixPath(rel));
-      }
-    }
-  };
-  await walk(texRoot);
-  return out;
-}
-
-/**
- * Copy the given texture from the cached client into the project directory and
- * update the project's metadata so it will be packaged.
- */
-export async function addTexture(
-  projectPath: string,
-  texture: string
-): Promise<void> {
-  const meta = await readProjectMeta(projectPath);
-  const root = await ensureAssets(meta.minecraft_version);
-  const src = path.join(root, 'assets', 'minecraft', 'textures', texture);
-  const dest = path.join(
-    projectPath,
-    'assets',
-    'minecraft',
-    'textures',
-    texture
-  );
-  await fs.promises.mkdir(path.dirname(dest), { recursive: true });
-  await fs.promises.copyFile(src, dest);
-  meta.assets.push(`assets/minecraft/textures/${texture}`);
-  await writeProjectMeta(projectPath, meta);
-}
-
-/**
- * Return the absolute path to a cached texture for the given project.
- */
-export async function getTexturePath(
-  projectPath: string,
-  texture: string
-): Promise<string> {
-  const meta = await readProjectMeta(projectPath);
-  const root = await ensureAssets(meta.minecraft_version);
-  return path.join(root, 'assets', 'minecraft', 'textures', texture);
-}
-
-/**
- * Convert a texture path into a URL that uses the custom `vanilla://` protocol
- * so the renderer can load the image without exposing file paths.
- */
-export async function getTextureURL(
-  projectPath: string,
-  texture: string
-): Promise<string> {
-  // Record paths used by the protocol so it can resolve this texture later
-  const meta = await readProjectMeta(projectPath);
-  const cacheRoot = await ensureAssets(meta.minecraft_version);
-  cacheTexturesDir = path.join(cacheRoot, 'assets', 'minecraft', 'textures');
-  return `vanilla://${texture}`;
-}
-
-/**
- * Combine multiple textures into a single PNG and return a data URL.
- */
-export async function createTextureAtlas(
-  projectPath: string,
-  textures: string[]
-): Promise<string> {
-  const meta = await readProjectMeta(projectPath);
-  const root = await ensureAssets(meta.minecraft_version);
-  const texDir = path.join(root, 'assets', 'minecraft', 'textures');
-
-  const composites: Array<{ input: Buffer; left: number; top: number }> = [];
-  let width = 0;
-  let height = 0;
-  for (const tex of textures) {
-    const img = sharp(path.join(texDir, tex));
-    const info = await img.metadata();
-    const buf = await img.png().toBuffer();
-    composites.push({ input: buf, left: width, top: 0 });
-    width += info.width ?? 0;
-    height = Math.max(height, info.height ?? 0);
-  }
-  if (width === 0 || height === 0) return '';
-  const out = await sharp({
-    create: {
-      width,
-      height,
-      channels: 4,
-      background: { r: 0, g: 0, b: 0, alpha: 0 },
-    },
-  })
-    .composite(composites)
-    .png()
-    .toBuffer();
-  return `data:image/png;base64,${out.toString('base64')}`;
-}
-
-/**
- * Register the `vanilla` protocol so it serves files directly from disk.
- */
-export function registerVanillaProtocol(protocol: Protocol) {
-  protocol.registerFileProtocol('vanilla', (request, callback) => {
-    const rel = decodeURI(request.url.replace('vanilla://', ''));
-    const cacheFile = cacheTexturesDir ? path.join(cacheTexturesDir, rel) : '';
-    callback(cacheFile);
-  });
-}
-
-/** Register the `asset` protocol to serve files from the active project. */
-export function registerAssetProtocol(protocol: Protocol) {
-  protocol.registerFileProtocol('asset', (request, callback) => {
-    const rel = decodeURI(request.url.replace('asset://', ''));
-    let file = activeProjectDir ? path.join(activeProjectDir, rel) : '';
-    if (/(\\|\/)$/.test(file)) {
-      file = file.substring(0, file.length - 1);
-    }
-    if (file.includes('?t=')) {
-      file = file.substring(0, file.indexOf('?t='));
-    }
-    callback(file);
-  });
-}
-
-/** Update the directories used by the custom protocols for the active project. */
-export async function setActiveProject(projectPath: string): Promise<void> {
-  const meta = await readProjectMeta(projectPath);
-  const cacheRoot = await ensureAssets(meta.minecraft_version);
-  cacheTexturesDir = path.join(cacheRoot, 'assets', 'minecraft', 'textures');
-  activeProjectDir = projectPath;
-}
-
-export function registerAssetHandlers(ipc: IpcMain) {
-  ipc.handle('add-texture', (_e, projectPath: string, texture: string) => {
-    return addTexture(projectPath, texture);
-  });
-
-  ipc.handle('list-textures', (_e, projectPath: string) => {
-    return listTextures(projectPath);
-  });
-
-  ipc.handle('get-texture-path', (_e, projectPath: string, tex: string) => {
-    return getTexturePath(projectPath, tex);
-  });
-
-  ipc.handle('get-texture-url', (_e, projectPath: string, tex: string) => {
-    return getTextureURL(projectPath, tex);
-  });
-
-  ipc.handle('create-atlas', (_e, projectPath: string, tex: string[]) => {
-    return createTextureAtlas(projectPath, tex);
-  });
-
-  ipc.handle('randomize-icon', (_e, projectPath: string) => {
-    return generatePackIcon(projectPath);
-  });
-}
