+++ src/main/projects.ts
@@ -5,7 +5,8 @@ import { dialog } from 'electron';
 import type { ProjectMetadata, PackMeta } from '../shared/project';
 import { PackMetaSchema } from '../shared/project';
 import { readProjectMeta, writeProjectMeta } from './projectMeta';
-import { listPackFormats, setActiveProject } from './assets';
+import { listPackFormats } from './assets';
+import { setActiveProject } from './assets/protocols';
 import { setLastProject } from './layout';
 import unzipper from 'unzipper';
 import { versionForFormat } from '../shared/packFormat';
